backend/requirements.txt
fastapi
uvicorn[standard]
gitpython
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
requests
